+++ src/main/scala/com/nulabinc/backlog/r2b/exporter/service/IssueInitializer.scala
@@ -1,9 +1,8 @@
 package com.nulabinc.backlog.r2b.exporter.service
 
-import java.io.{File, FileOutputStream}
-import java.net.{HttpURLConnection, URL}
-import java.nio.channels.Channels
+import java.net.URL
 
+import better.files.{File => Path}
 import com.nulabinc.backlog.migration.common.convert.Convert
 import com.nulabinc.backlog.migration.common.domain._
 import com.nulabinc.backlog.migration.common.utils.{DateUtil, IOUtil, Logging, StringUtil}
@@ -12,8 +11,6 @@ import com.nulabinc.backlog.r2b.redmine.conf.RedmineConstantValue
 import com.nulabinc.backlog.r2b.redmine.domain.RedmineCustomFieldDefinition
 import com.nulabinc.backlog.r2b.utils.TextileUtil
 import com.taskadapter.redmineapi.bean._
-import better.files.{File => Path}
-import com.nulabinc.backlog.migration.common.utils.ControlUtil.using
 
 import scala.jdk.CollectionConverters._
 
@@ -219,46 +216,12 @@ private[exporter] class IssueInitializer(exportContext: ExportContext, issueDirP
   }
 
   private[this] def attachment(attachment: Attachment): Unit = {
-
     val dir  = exportContext.backlogPaths.issueAttachmentDirectoryPath(issueDirPath)
     val path = exportContext.backlogPaths.issueAttachmentPath(dir, attachment.getFileName)
     IOUtil.createDirectory(dir)
 
     val url: URL = new URL(s"${attachment.getContentURL}?key=${exportContext.apiConfig.key}")
-    val redirected = followRedirect(url)
 
-    doDownload(redirected, path.path.toFile)
+    AttachmentService.download(url, path.path.toFile)
   }
-
-  private val MAX_REDIRECT_COUNT = 10
-
-  private[this] def followRedirect(url: URL, count: Int = 0): URL =
-    url.openConnection match {
-      case http: HttpURLConnection =>
-        http.setRequestMethod("GET")
-        http.connect()
-        using(http) { connection =>
-          connection.getResponseCode match {
-            case 302 | 303 =>
-              val newUrl = new URL(connection.getHeaderField("Location"))
-              if (count < MAX_REDIRECT_COUNT) followRedirect(newUrl, count + 1) else newUrl
-            case _ =>
-              url
-          }
-        }
-      case _ => url
-    }
-
-  private[this] def doDownload(url: URL, file: File): Unit =
-    try {
-      val rbc = Channels.newChannel(url.openStream())
-      val fos = new FileOutputStream(file)
-      fos.getChannel.transferFrom(rbc, 0, java.lang.Long.MAX_VALUE)
-
-      rbc.close()
-      fos.close()
-    } catch {
-      case e: Throwable => logger.warn("Download issue attachment failed: " + e.getMessage)
-    }
-
 }
